+++ R/Tools4CommonMetabolomics.R
@@ -1,7 +1,6 @@
 ################################################################################
 # Adduct mz calculation --------------------------------------------------------
-
-# ca
+#   calculateExactMass ---------------------------------------------------------
 #' @title calculateExactMass
 #' @author Zhiwei Zhou
 #' @description calculate exact mass from formula
@@ -22,6 +21,7 @@ setGeneric(name = 'calculateExactMass',
            })
 
 
+#   calculateMz ----------------------------------------------------------------
 #' @title calculateMz
 #' @author Zhiwei Zhou
 #' @description calculate adduct mz with exact mass
@@ -60,6 +60,20 @@ setGeneric(name = 'calculateMz',
              mz
            })
 
+
+#   transformMz ----------------------------------------------------------------
+#' @title transformMz
+#' @author Zhiwei Zhou
+#' @param exact_mass numeric
+#' @param formula Default: NULL
+#' @param adduct_list Default: NULL
+#' @param type 'adduct' or 'nl'.
+#' @param polarity 'positive' or 'negative'
+#' @export
+#' @examples
+#' transformMz(exact_mass = 180.0634, type = 'adduct', polarity = 'positive')
+#' transformMz(exact_mass = 180.0634, type = 'nl', polarity = 'positive')
+#' transformMz(exact_mass = 180.0634, adduct_list = c('[M-H]-', '[M+H]+'))
 
 # exact_mass <- 180.0634
 # adduct <- '[M-H]-'
@@ -140,6 +154,31 @@ setGeneric(name = 'transformMz',
 
 
 
+#' @title convertMz2Adduct
+#' @author Zhiwei ZHou
+#' @description mz conversion between different adduct forms
+#' @param base_mz numeric
+#' @param base_adduct character
+#' @param ... Other arguments passed on to [transformMz()]
+#' @export
+#' @examples
+#' convertMz2Adduct(base_mz = 181.0707,
+#'                  base_adduct = '[M+H]+',
+#'                  adduct_list = NULL,
+#'                  type = 'adduct',
+#'                  polarity = 'positive')
+#'
+#' convertMz2Adduct(base_mz = 181.0707,
+#'                  base_adduct = '[M+H]+',
+#'                  adduct_list = c('[M-H2O+H]+', '[M+Na]+'),
+#'                  type = 'nl',
+#'                  polarity = 'negative')
+#'
+#' convertMz2Adduct(base_mz = 143.0347,
+#'                  base_adduct = '[3M-H]-',
+#'                  adduct_list = c('[M-H2O+H]+', '[M+Na]+'))
+
+
 # convertMz2Adduct(base_mz = 181.0707,
 #                  base_adduct = '[M+H]+',
 #                  adduct_list = NULL,
@@ -208,19 +247,20 @@ setGeneric(name = 'convertMz2Adduct',
 
 
 
-
-#' @title generateFormulaMass
+#   generateFormulaMassFromSmiles ----------------------------------------------
+#' @title generateFormulaMassFromSmiles
 #' @author Zhiwei Zhou
+#' @description generate formula and mass from smiles
 #' @param smiles
 #' @return a data.frame with exact_mass and formula
 #' @export
 #' @examples
-#' test <- generateFormulaMass('CN1C=NC(C[C@H](N)C(O)=O)=C1')
+#' test <- generateFormulaMassFromSmiles('CN1C=NC(C[C@H](N)C(O)=O)=C1')
 
 
-# generateFormulaMass('CN1C=NC(C[C@H](N)C(O)=O)=C1')
+# generateFormulaMassFromSmiles('CN1C=NC(C[C@H](N)C(O)=O)=C1')
 
-setGeneric('generateFormulaMass',
+setGeneric('generateFormulaMassFromSmiles',
            def = function(smiles) {
              if (is.na(smiles) | is.null(smiles)) {
                formula_result <- data.frame(exact_mass=NA,
@@ -250,3 +290,133 @@ setGeneric('generateFormulaMass',
            }
 )
 
+
+
+
+################################################################################
+# structure format conversion --------------------------------------------------
+#   convertStructureFormat -----------------------------------------------------
+
+#' @title convertStructureFormat
+#' @author Zhiwei Zhou
+#' @description R wrapper for OpebBabel
+#' @param file_input
+#' @param file_output
+#' @param format_input See \link{https://open-babel.readthedocs.io/en/latest/FileFormats/Overview.html#file-formats}
+#' @param format_output See \link{https://open-babel.readthedocs.io/en/latest/FileFormats/Overview.html#file-formats}
+#' @param options
+#' @export
+#' @example
+#' file_input <- 'I:/00_projects/03_MetDNA2/00_data/20200706_KEGG_settle_and_remove_redunancy/test.smiles'
+#' file_output <- 'I:/00_projects/03_MetDNA2/00_data/20200706_KEGG_settle_and_remove_redunancy/test_output.inchikey'
+#' format_input <- 'smiles'
+#' format_output <- 'inchikey'
+
+
+# file_input <- 'I:/00_projects/03_MetDNA2/00_data/20200706_KEGG_settle_and_remove_redunancy/test.smiles'
+# file_output <- 'I:/00_projects/03_MetDNA2/00_data/20200706_KEGG_settle_and_remove_redunancy/test_output.inchikey'
+# format_input <- 'smiles'
+# format_output <- 'inchikey'
+# convertStructureFormat(file_input = file_input,
+#                        file_output = file_output)
+
+setGeneric(name = 'convertStructureFormat',
+           def = function(
+             file_input,
+             file_output,
+             format_input,
+             format_output,
+             options = ''
+           ){
+             command <- paste0('obabel ',
+                               '-i',
+                               format_input,
+                               ' ',
+                               file_input,
+                               ' -o',
+                               format_output,
+                               ' -O',
+                               file_output)
+
+             if (nchar(options) != 0) {
+               paste0(command, ' ', options)
+             }
+
+             system(command = command)
+
+           }
+)
+
+
+################################################################################
+# formula functions ------------------------------------------------------------
+
+#' @title limitElements
+#' @author Zhiwei Zhou
+#' @param cpd_formula
+#' @param element_necessary Necessary elements. Default: 'C'
+#' @param element_included Limit elements. Default: c("C", 'H', 'N', 'O', 'P', 'S', 'F', 'Cl', 'Br', 'I')
+#' @return Logistical value. TRUE or FALSE
+#' @export
+#' @example
+#' cpd_formula <- 'H2O'
+#' limitElements(cpd_formula)
+#' cpd_formula <- 'C6H12O6'
+#' limitElements(cpd_formula)
+#' cpd_formula <- 'C55H68MgN4O6'
+#' limitElements(cpd_formula)
+
+# cpd_formula <- 'H2O'
+# limitElements(cpd_formula)
+# cpd_formula <- 'C6H12O6'
+# limitElements(cpd_formula)
+# cpd_formula <- 'C55H68MgN4O6'
+# limitElements(cpd_formula)
+# cpd_formula <- 'C53H60ClN3O16'
+# limitElements(cpd_formula)
+
+
+setGeneric(name = 'limitElements',
+           def = function(cpd_formula,
+                          element_necessary = 'C',
+                          element_included = c("C", 'H', 'N', 'O', 'P', 'S', 'F', 'Cl', 'Br', 'I')){
+             if (missing(cpd_formula)) {
+               stop('Please input cpd_formula')
+             }
+
+             element_composition <- try(CHNOSZ::makeup(formula = cpd_formula),
+                                        silent = TRUE)
+
+
+
+             element_all <- names(element_composition)
+
+             if (!(element_necessary %in% element_all)) {
+               return(FALSE)
+             }
+
+             if (all(element_all %in% element_included)) {
+               return(TRUE)
+             } else {
+               return(FALSE)
+             }
+
+           })
+
+
+################################################################################
+# startup message --------------------------------------------------------------
+.onAttach <- function(libname, pkgname){
+  packageStartupMessage("
+If you have any questions, please send email to zhiwei92@126.com.
+Authors: Zhiwei Zhou.
+Maintainer: Zhiwei Zhou
+Version 0.0.1.01 (20210125)
+--------------
+* Add module: Tools4PublicDB
+* Add functions: extractPropertyFromHMDB(), extractPropertyFromPubChem(), convertStructureFormat()
+* Add manual for: calculateMz, transformMz, generateFormulaMassFromSmiles
+* Add function name: generateFormula ---> generateFormulaMassFromSmiles
+* Add functions: limitElements()
+")
+}