+++ NEWS.md
@@ -4,3 +4,14 @@
 
 # 20210107 v0.0.0.9001
 * Added generateFormulaMass(), convertAdduct4Tools() functions
+
+
+# 20210114 v0.0.1.0
+* Add module: Tools4PublicDB
+* Add functions: extractPropertyFromHMDB(), extractPropertyFromPubChem(), convertStructureFormat()
+* Add manual for: calculateMz, transformMz, generateFormulaMassFromSmiles
+* Add function name: generateFormula ---> generateFormulaMassFromSmiles
+
+
+# 20210114 v0.0.1.01
+* Add functions: limitElements()